+++ ps2_cam_acolyte/games/shared_ui.h
@@ -19,6 +19,13 @@ public:
 		ImGui::TextColored(toggle.currently_tweaking() ? ui_colors::on_obvious : ui_colors::off_obvious, toggle.currently_tweaking() ? "ON" : "OFF");
 	}
 
+	template <size_t N>
+	static void toggle(const controller& c, button_type button, const restoring_toggle_state<N>& toggle, const char* label)
+	{
+		shared_ui::button(c, button); ImGui::SameLine(); ImGui::Text("%s: ", label); ImGui::SameLine();
+		ImGui::TextColored(toggle.is_on() ? ui_colors::on_obvious : ui_colors::off_obvious, toggle.is_on() ? "ON" : "OFF");
+	}
+
 	static void button(const controller& c, button_type button)
 	{
 		ImGui::Text("[%s]", c.get_button_display_name(button));