+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+**v.02**
+
+Added support for:
+
+| UUID     | Game                                | Notes                                                        |
+| -------- | ----------------------------------- | ------------------------------------------------------------ |
+| 901AAC09 | Haunting Ground (USA)                 |                                                              |
+| 9AC63A2E | Kuon (USA)                            |                                                              |
+| 053D2239 | Metal Gear Solid 3: Subsistence (USA) |                                                              |
+| D6C48447 | Siren (USA)                           |                                                              |
+
+- There is now a simple record & playback system for camera movements, with speed and smoothing controls.
+- Proper button names(i.e. Cross vs. A) are now displayed for common controller types.
+- Fixed a issue that could crash PCSX2 when turning off the free cam for Silent Hill Origins.
+
 **v0.01**
 
 Initial release with support for: